refactor: Рефакторинг propose-idea страницы
