+++ src/app/propose-idea/stepDataEn.json
@@ -4,39 +4,33 @@
     "header": "Project Creation",
     "title": "What needs to be done?",
     "fieldName": "description",
+    "isMultiline": true,
     "labelTitle": "Describe what needs to be done",
     "showExample": "Show example",
     "exampleText": "A service where I can send a link to a track from any streaming service. In response, a list of links to the same track on all possible streaming platforms, including YouTube, is generated.",
-    "buttons": [
-      {"Cancel": "Cancel"},
-      {"Next": "Next"}
-    ]
+    "buttons": [{ "Cancel": "Cancel" }, { "Next": "Next" }]
   },
   {
     "label": "STEP 2",
     "header": "Project Creation",
     "title": "Who is Your Client",
     "fieldName": "additional",
+    "isMultiline": true,
     "labelTitle": "Describe your clients",
     "showExample": "Show example",
     "exampleText": "Example text for STEP 2.",
-    "buttons": [
-      {"Back": "Back"},
-      {"Next": "Next"}
-    ]
+    "buttons": [{ "Back": "Back" }, { "Next": "Next" }]
   },
   {
     "label": "STEP 3",
     "header": "Project Creation",
     "title": "Specify the Project Name",
     "fieldName": "title",
+    "isMultiline": false,
     "labelTitle": "Project Name",
     "showExample": "Show example",
     "exampleText": "Example text for STEP 3.",
-    "buttons": [
-      {"Back": "Back"},
-      {"CreateProject": "Create Project"}
-    ]
+    "buttons": [{ "Back": "Back" }, { "CreateProject": "Create Project" }]
   },
   {
     "label": "STEP 4",
@@ -47,9 +41,9 @@
     "projectTitle": "Project Name:",
     "isCorrectTitle": "Is everything correct?",
     "buttons": [
-      {"EditData": "Edit Data"},
-      {"Cancel": "Cancel"},
-      {"Approve": "Yes, all correct"}
+      { "EditData": "Edit Data" },
+      { "Cancel": "Cancel" },
+      { "Approve": "Yes, all correct" }
     ]
   }
 ]
