+++ src/app/propose-idea/stepDataRu.json
@@ -4,39 +4,33 @@
     "header": "Создание проекта",
     "title": "Что нужно сделать?",
     "fieldName": "description",
+    "isMultiline": true,
     "labelTitle": "Опишите, что нужно сделать",
     "showExample": "Показать пример",
     "exampleText": "Сервис, в который я могу отправить ссылку на трек из любого стримингового сервиса. В ответ генерируется список ссылок на этот же трек на всех возможных стриминговых площадках, включая YouTube.",
-    "buttons": [
-      {"Cancel": "Отмена"},
-      {"Next":  "Далее"}
-    ]
+    "buttons": [{ "Cancel": "Отмена" }, { "Next": "Далее" }]
   },
   {
     "label": "ЭТАП 2",
     "header": "Создание проекта",
     "title": "Кто ваш клиент",
     "fieldName": "additional",
+    "isMultiline": true,
     "labelTitle": "Опишите, ваших клиентов",
     "showExample": "Показать пример",
     "exampleText": "Пример текста для этапа 2.",
-    "buttons": [
-      {"Back": "Назад"},
-      {"Next":  "Далее"}
-    ]
+    "buttons": [{ "Back": "Назад" }, { "Next": "Далее" }]
   },
   {
     "label": "ЭТАП 3",
     "header": "Создание проекта",
     "title": "Укажите название проекта",
     "fieldName": "title",
+    "isMultiline": false,
     "labelTitle": "Название проекта",
     "showExample": "Показать пример",
     "exampleText": "Пример текста для этапа 3.",
-    "buttons": [
-      {"Back": "Назад"},
-      {"CreateProject":  "Создать проект"}
-    ]
+    "buttons": [{ "Back": "Назад" }, { "CreateProject": "Создать проект" }]
   },
   {
     "label": "ЭТАП 4",
@@ -47,9 +41,9 @@
     "projectTitle": "Название проекта:",
     "isCorrectTitle": "Все верно?",
     "buttons": [
-      {"EditData": "Изменить данные"},
-      {"Cancel":  "Отмена"},
-      {"Approve":  "Да, всё верно"}
+      { "EditData": "Изменить данные" },
+      { "Cancel": "Отмена" },
+      { "Approve": "Да, всё верно" }
     ]
   }
 ]
